requirements.txt
python-telegram-bot==20.7
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests
